[BROKEN] Test utils/index.ts
The goal here is to have test coverage on all the functions that use
Moment, so that when we switch to another library we can be confident
that all of them still work as expected.

However, I am having trouble getting Jest to work with this project.

Saving these changes and coming back to them later, after other sprint
work is complete.