Fix height calculation for work history column
When calculating the number of rows for this grid, the browser uses
approximately equally-heighted rows for each (why?). This is somehow
related to the `grid-row: span 2` setting on this right-side column.

Using a larger span setting (in this case 20) means that each of the
first two _real_ rows can match the height of the items on the left, and
the excessive height of the right-side column spills into the
superfluous rows.

I find this solution deeply unsatisfying, and wish I could answer that
"why" question above instead.