Improve row height math
This improves #83 / 6c41a48. Rather than use a hack on the
`grid-row: span X` for the right-side item, this instructs the browser
to auto-size the first row to the content, and use the rest of the
available space for the second row. This means the first row will be as
short as possible.

The screenshots from #83 are still relevant here; this is an internal
logic change only, which I believe makes the CSS more understandable.